+++ src/main/java/com/spring/timebook/config/properties/OAuthKakaoProperty.java
@@ -1,4 +1,4 @@
-package com.spring.timebook.config;
+package com.spring.timebook.config.properties;
 
 
 import lombok.Getter;
@@ -8,7 +8,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "login.kakao")
 @Getter
 @Setter
-public class OAuthKakaoConfig {
+public class OAuthKakaoProperty {
     private String apiKey;
     private String redirectUri;
 }
