+++ src/main/java/com/spring/timebook/config/properties/OAuthNaverProperty.java
@@ -1,5 +1,4 @@
-package com.spring.timebook.config;
-
+package com.spring.timebook.config.properties;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,7 +7,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "login.naver")
 @Getter
 @Setter
-public class OAuthNaverConfig {
+public class OAuthNaverProperty {
     private String clientId;
     private String clientSecret;
     private String redirectUri;
